consolidate useLocation code
